crimson-iq/deployment/dockerfiles/dockerfiles/collector/requirements.txt
paho-mqtt==1.6.1
pymongo==4.7.2